Add level to antiremoval on overlay Add tests
