+++ url-gen/src/test/kotlin/com/cloudinary/transformation/TransformationTest.kt
@@ -45,16 +45,16 @@ class TransformationTest {
             })
     }
 
-    @Test
-    fun testAntiRemoval() {
-        cldAssert("l_sample/e_anti_removal,fl_layer_apply", Transformation().antiRemoval(imageSource))
-        cldAssert(
-            "l_sample/e_sepia/e_anti_removal,fl_layer_apply",
-            Transformation().antiRemoval(AntiRemoval.source(image("sample") {
-                transformation(sepiaTransformation)
-            }))
-        )
-    }
+//    @Test
+//    fun testAntiRemoval() {
+//        cldAssert("l_sample/e_anti_removal,fl_layer_apply", Transformation().antiRemoval(imageSource))
+//        cldAssert(
+//            "l_sample/e_sepia/e_anti_removal,fl_layer_apply",
+//            Transformation().antiRemoval(AntiRemoval.source(image("sample") {
+//                transformation(sepiaTransformation)
+//            }))
+//        )
+//    }
 
     @Test
     fun testCutout() {
@@ -270,7 +270,7 @@ class TransformationTest {
                 overlay(Overlay.image {
                     source("sample") {
 
-                    }.blendMode(BlendMode.antiRemoval(15))
+                    }.blendMode(BlendMode.antiRemoval { level(15) })
                 })
 
             }
